Add GitHub-inspired design system and implement simple robot runner game
- Introduced a new CSS file for a GitHub-inspired design system, defining variables for colors, typography, spacing, and component styles.
- Implemented a simple robot runner game in JavaScript, featuring basic mechanics such as jumping, obstacle generation, collision detection, and score tracking.
- Enhanced user experience with responsive design and animations for game elements.
